+++ config/projects/gitlab.rb
@@ -97,6 +97,7 @@ dependency "remote-syslog" if ee
 dependency "logrotate"
 dependency "runit"
 dependency "nodejs"
+dependency "kubectl"
 dependency "gitlab-rails"
 dependency "gitlab-shell"
 dependency "gitlab-workhorse"