+++ CHANGELOG.md
@@ -7,6 +7,7 @@ omnibus-gitlab repository.
 
 - Add support for registry debug addr configuration
 - Support specifying a post reconfigure script to run in the docker container
+- Allow users to specify an initial shared runner registration token
 - Updated cacerts.pem to 2016-09-14 version
 
 8.12.3