+++ CHANGELOG.md
@@ -6,6 +6,7 @@ omnibus-gitlab repository.
 8.13.0
 
 - Add support for registry debug addr configuration
+- Allow users to specify an initial shared runner registration token
 - Add support for configuring workhorse's api limiting
 - Update the storage directory helper to check permissions for symlink targets
 - Support specifying a post reconfigure script to run in the docker container