+++ CHANGELOG.md
@@ -22,7 +22,7 @@ omnibus-gitlab repository.
 8.12.3
 
 - Updated cacerts.pem to 2016-09-14 version
-
+.
 8.12.2
 
 - Update openssl to 1.0.2j