+++ CHANGELOG.md
@@ -6,9 +6,19 @@ omnibus-gitlab repository.
 8.13.0
 
 - Add support for registry debug addr configuration
+- Allow users to specify an initial shared runner registration token
+- Add support for configuring workhorse's api limiting
+- Update the storage directory helper to check permissions for symlink targets
 - Support specifying a post reconfigure script to run in the docker container
 - Allow users to specify an initial shared runner registration token
 - Updated cacerts.pem to 2016-09-14 version
+- Add support for nginx status
+- Enable jemalloc by default 0a7799d2
+- Move database migration log to a persisted location
+
+8.12.4
+
+- No changes
 
 8.12.3